requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
transformers>=4.21.0
torch>=1.12.0
accelerate>=0.20.0
